requirements.txt
pandas
Pillow
python-dotenv
google-generativeai
openai
streamlit